Time: 3 ms (54.95%), Space: 62.4 MB (26.37%) - LeetHub

+++ 2874-maximum-value-of-an-ordered-triplet-ii/2874-maximum-value-of-an-ordered-triplet-ii.java
@@ -13,7 +13,7 @@ class Solution {
         }
         long res=0;
         for(int i=1;i<len-1;i++){
-            res=Math.max(res,(lg[i]-nums[i])*rg[i]);
+            res=Math.max(res,(long)(lg[i]-nums[i])*rg[i]);
         }
         return res;
     }
